+++ Source/Core/Entities/EntityLifecycleState.cs
@@ -1,0 +1,9 @@
+namespace Mythos.Framework.Entities;
+
+public enum EntityLifecycleState
+{
+    Active,
+    Inactive,
+    Retired,
+    Destroyed,
+}